+++ test/models/diffusion/test_song_unet_pos_lt_embd.py
@@ -354,12 +354,28 @@ def test_song_unet_optims(device):
     # Check JIT
     model, invar = setup_model()
     assert common.validate_jit(model, (*invar,))
-    # Check AMP
+    # Check AMP with amp_mode=True for the layers: should pass
     model, invar = setup_model()
+    model.amp_mode = True
     assert common.validate_amp(model, (*invar,))
-    # Check Combo
+    # Check Combo with amp_mode=True for the layers: should pass
     model, invar = setup_model()
+    model.amp_mode = True
     assert common.validate_combo_optims(model, (*invar,))
+
+    # Check failures (only on GPU, because validate_amp and validate_combo_optims
+    # don't activate amp for SongUNetPosLtEmbd on CPU)
+    if device == "cuda:0":
+        # Check AMP: should fail because amp_mode is False for the layers
+        with pytest.raises(RuntimeError):
+            model, invar = setup_model()
+            assert common.validate_amp(model, (*invar,))
+        # Check Combo: should fail because amp_mode is False for the layers
+        # NOTE: this test doesn't fail because validate_combo_optims doesn't
+        # activate amp for SongUNetPosLtEmbd, even on GPU
+        # with pytest.raises(RuntimeError):
+        #     model, invar = setup_model()
+        #     assert common.validate_combo_optims(model, (*invar,))
 
 
 # Skip CPU tests because too slow